Added LocationUtils (#44)
